Move encoder and decoder functions into sketchlib
