+++ sketchlib/dx7/decode_dx7.js
@@ -8,14 +8,17 @@ import {
   SUB_STATUS,
   VOICE_LENGTH,
   VOICE_START,
-} from "../sketchlib/dx7/dx7_constants.js";
-import { DX7Cartridge } from "./DX7Cartridge.js";
-import { DX7Envelope } from "./DX7Envelope.js";
-import { DX7FreqSettings } from "./DX7FreqSettings.js";
-import { DX7KeyLevelScaling, DX7ScalingCurve } from "./DX7KeyLevelScaling.js";
-import { DX7LFO } from "./DX7LFO.js";
-import { DX7Operator } from "./DX7Operator.js";
-import { DX7Voice } from "./DX7Voice.js";
+} from "./dx7_constants.js";
+import { DX7Cartridge } from "../../DX7PatchViewer/DX7Cartridge.js";
+import { DX7Envelope } from "../../DX7PatchViewer/DX7Envelope.js";
+import { DX7FreqSettings } from "../../DX7PatchViewer/DX7FreqSettings.js";
+import {
+  DX7KeyLevelScaling,
+  DX7ScalingCurve,
+} from "../../DX7PatchViewer/DX7KeyLevelScaling.js";
+import { DX7LFO } from "../../DX7PatchViewer/DX7LFO.js";
+import { DX7Operator } from "../../DX7PatchViewer/DX7Operator.js";
+import { DX7Voice } from "../../DX7PatchViewer/DX7Voice.js";
 
 /**
  * Decode an envelope
